Refactor remote attestation
