+++ remote_attestation/rust/Cargo.toml
@@ -6,15 +6,16 @@ edition = "2021"
 license = "Apache-2.0"
 
 [features]
-default = []
+default = ["ring-crypto"]
 std = ["anyhow/std", "prost/std"]
+ring-crypto = ["ring"]
 
 [dependencies]
 anyhow = { version = "*", default-features = false }
 bytes = { version = "*", default-features = false }
 log = "*"
 prost = { version = "*", default-features = false, features = ["prost-derive"] }
-ring = { path = "../../third_party/ring", default-features = false }
+ring = { path = "../../third_party/ring", default-features = false, optional = true }
 
 [build-dependencies]
 prost-build = "*"
